move from enterprise/ and fix tests
