+++ web/src/views/QueryInputInViewContent.tsx
